requirements.txt
Pillow==2.0.0
